Create README.md
In this game, the user is given 9 chances to guess the letter the computer is guessing. The computer picks a random letter from an array of each letter in the alphabet. The user is asked to input a key from a key listener after each guess. If the user guess matches
the computer guess, then a win is recorded. If the user guess does not match the computer guess, nothing happens. However, if the user is out of guesses, then a loss is recorded.